requirements.txt
fastapi>=0.112,<1.0
uvicorn[standard]>=0.29
gunicorn>=21.2
pydantic>=2.6
pydantic-settings>=2.4
python-dotenv>=1.0
hdbcli